+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 == HEAD
 
+== 1.0.3 (May 7, 2013)
+
 * Fix AMD export modification made in v1.0.1.
 * Only add orphan elements to a `documentFragment` when necessary.
 
-== 1.0.2 (May 7, 20313)
+== 1.0.2 (May 7, 2013)
 
 * Fix `matches()` on orphan elements in IE 9.